fix: resolve all 485 mypy type errors across 44 source files
- estimators/: 121 errors (type narrowing, Optional handling, return types)
- policy/: 117 errors (floating type params, ndarray/list annotations)
- discovery/ + interference/: 99 errors (Optional narrowing, dict typing)
- visualization/ + sensitivity/: 86 errors (matplotlib types, Optional narrowing)
- api/ + ml/ + misc: 62 errors (union-attr, import stubs, Python 3.9 compat)

Common fix patterns applied:
- Add None checks / assertions for Optional type narrowing
- Wrap numpy returns with np.asarray() or float() for no-any-return
- Add missing generic type parameters (np.floating[Any], Callable[..., Any])
- Fix implicit Optional (PEP 484) with explicit Optional[X]
- Add type: ignore for untyped third-party imports (plotly, cvxpy, psutil)
- Use from __future__ import annotations for Python 3.9 X|Y syntax compat

Zero runtime behavior changes. mypy now reports 0 errors across 114 files.

Closes #170, closes #171, closes #172, closes #173, closes #174

+++ libs/causal_inference/causal_inference/evaluation/hte_metrics.py
@@ -63,7 +63,7 @@ def pehe_score(
         Journal of Computational and Graphical Statistics, 20(1), 217-240.
     """
     mse = mean_squared_error(y_true, y_pred)
-    return mse if squared else np.sqrt(mse)
+    return float(mse) if squared else float(np.sqrt(mse))
 
 
 def policy_value(
